[FE] refactor #73: 리뷰 페이지 구조 수정
